Refactor sign-in/sign-up pages
