+++ src/components/Footer.astro
@@ -4,7 +4,7 @@ import ThemeSelect from "@/components/ThemeSelect.svelte";
 
 <footer class="border-t border-border">
   <div
-    class="mx-auto max-w-7xl px-6 py-12 md:flex md:items-center md:justify-between lg:px-8"
+    class="mx-auto max-w-7xl px-6 py-4 md:flex md:items-center md:justify-between lg:px-8"
   >
     <div class="flex justify-center gap-x-6 md:order-2">
       <ThemeSelect client:load />
