--- src/components/Logo.svelte
@@ -1,32 +1,0 @@
-<script>
-  export let className = "size-8";
-</script>
-
-<svg
-  xmlns="http://www.w3.org/2000/svg"
-  fill="none"
-  viewBox="0 0 24 24"
-  strokeWidth="1.5"
-  class={className}
->
-  <defs>
-    <linearGradient id="gradient" x1="0%" y1="0%" x2="100%" y2="100%">
-      <stop offset="0%" style="stop-color: #4f46e5" />
-      <stop offset="0%" style="stop-color: #e11d48" />
-      <stop offset="100%" style="stop-color: #0ea5e9" />
-    </linearGradient>
-  </defs>
-
-  <defs>
-    <filter id="vibrancy">
-      <feColorMatrix type="saturate" values="5" />
-    </filter>
-  </defs>
-  <path
-    stroke="url(#gradient)"
-    stroke-linecap="round"
-    stroke-linejoin="round"
-    d="M15.59 14.37a6 6 0 0 1-5.84 7.38v-4.8m5.84-2.58a14.98 14.98 0 0 0 6.16-12.12A14.98 14.98 0 0 0 9.631 8.41m5.96 5.96a14.926 14.926 0 0 1-5.841 2.58m-.119-8.54a6 6 0 0 0-7.381 5.84h4.8m2.581-5.84a14.927 14.927 0 0 0-2.58 5.84m2.699 2.7c-.103.021-.207.041-.311.06a15.09 15.09 0 0 1-2.448-2.448 14.9 14.9 0 0 1 .06-.312m-2.24 2.39a4.493 4.493 0 0 0-1.757 4.306 4.493 4.493 0 0 0 4.306-1.758M16.5 9a1.5 1.5 0 1 1-3 0 1.5 1.5 0 0 1 3 0Z"
-    filter="url(#vibrancy)"
-  />
-</svg>
