+++ src/pages/sign-in.astro
@@ -1,0 +1,113 @@
+---
+if (Astro.locals.user?.id) return Astro.redirect("/");
+import Layout from "@/layouts/Layout.astro";
+import {Button} from "@/components/primitives/button";
+import {Input} from "@/components/primitives/input";
+import PasswordField from "@/components/PasswordField.svelte";
+import Mail from "@/assets/mail.svg";
+import {authClient} from "@/lib/auth-client";
+
+const errors = {
+  email: "",
+  password: "",
+  server: "",
+};
+const prefilled = {
+  email: "",
+  password: "",
+};
+
+if (Astro.request.method === "POST") {
+  try {
+    const formData = await Astro.request.formData();
+    const email = formData.get("email") as string;
+    const password = formData.get("password") as string;
+    prefilled.email = email;
+    prefilled.password = password;
+
+    if (!email || !password) {
+      errors.email = "Email is required";
+      errors.password = "Password is required";
+      return;
+    }
+
+    const result = await authClient.signIn.email({
+      email,
+      password,
+    });
+
+    console.log(result.error);
+
+    if (Boolean(result.error) === false) {
+      return Astro.redirect("/dashboard");
+    } else {
+      if (result.error?.code === "invalid_credentials") {
+        errors.server = "Invalid credentials";
+      } else {
+        errors.server =
+          result.error?.message || "An error occurred during sign in";
+      }
+    }
+  } catch (error) {
+    console.error(error);
+    errors.server = "Something went wrong: " + error;
+  }
+}
+---
+
+<Layout>
+  <div
+    class="flex flex-col items-center justify-center w-full my-12 min-h-[70dvh]"
+  >
+    <div class="w-full max-w-xs">
+      <h1 class="text-2xl font-bold mb-8 text-center">Sign In</h1>
+
+      <form method="POST" class="space-y-4">
+        <div
+          class="relative flex items-center rounded-md border focus-within:ring-1 focus-within:ring-ring pl-2"
+        >
+          <Mail class="h-5 w-5 text-muted-foreground" />
+          <Input
+            type="email"
+            placeholder="Email"
+            class="border-0 focus-visible:ring-0 shadow-none bg-transparent dark:bg-transparent"
+            name="email"
+            value={prefilled.email}
+            required
+          />
+          {
+            errors.email && (
+              <span class="text-red-500 text-xs ml-2">{errors.email}</span>
+            )
+          }
+        </div>
+
+        <PasswordField
+          name="password"
+          required
+          client:load
+          value={prefilled.password}
+        />
+
+        {
+          errors.password && (
+            <span class="text-red-500 text-xs ml-2">{errors.password}</span>
+          )
+        }
+
+        {
+          errors.server && (
+            <span class="text-red-500 text-xs ml-2">{errors.server}</span>
+          )
+        }
+
+        <Button type="submit" class="w-full mt-4">Sign In</Button>
+
+        <p class="text-center text-sm mt-4">
+          Don't have an account?{" "}
+          <a href="/sign-up" class="text-primary underline">Sign up here</a>
+        </p>
+      </form>
+    </div>
+  </div>
+</Layout>
