+++ src/pages/sign-up.astro
